requirements.txt
# requirements.txt
streamlit>=1.32
requests>=2.31
PyPDF2>=3.0
pdf2image>=1.16
pytesseract>=0.3.10
python-dotenv>=1.0
pyyaml>=6.0
tqdm>=4.66
python-multipart>=0.0.6